+++ tests/SequenceTest.php
@@ -241,6 +241,23 @@ class SequenceTest extends TestCase
         $this->assertSame([2, 4], $b->toList());
     }
 
+    public function testExclude()
+    {
+        $a = Sequence::of()
+            ->add(1)
+            ->add(2)
+            ->add(3)
+            ->add(4);
+        $b = $a->exclude(static function(int $value): bool {
+            return $value % 2 === 0;
+        });
+
+        $this->assertInstanceOf(Sequence::class, $b);
+        $this->assertNotSame($a, $b);
+        $this->assertSame([1, 2, 3, 4], $a->toList());
+        $this->assertSame([1, 3], $b->toList());
+    }
+
     public function testForeach()
     {
         $sum = 0;