+++ tests/SetTest.php
@@ -227,6 +227,23 @@ class SetTest extends TestCase
         $this->assertSame([2, 4], $s2->toList());
     }
 
+    public function testExclude()
+    {
+        $s = Set::of()
+            ->add(1)
+            ->add(2)
+            ->add(3)
+            ->add(4);
+
+        $s2 = $s->exclude(static function(int $v) {
+            return $v % 2 === 0;
+        });
+        $this->assertNotSame($s, $s2);
+        $this->assertInstanceOf(Set::class, $s2);
+        $this->assertSame([1, 2, 3, 4], $s->toList());
+        $this->assertSame([1, 3], $s2->toList());
+    }
+
     public function testForeach()
     {
         $s = Set::of()